rename to sanity.config
